Unify request utilities
